requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.6
djangorestframework==3.15.2
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
PyMuPDF==1.25.3
pypdfium2==4.30.1
pytesseract==0.3.13
sqlparse==0.5.3